+++ R/vector.R
@@ -6,3 +6,7 @@ renv_vector_diff <- function(x, y) {
 renv_vector_intersect <- function(x, y) {
   y[match(x, y, 0L)]
 }
+
+renv_vector_unique <- function(x) {
+  x[!duplicated(x)]
+}